Convert swedish-crowns to typescript
Fix types for swedish-crowns
